+++ SmartSaver.WebApi/Controllers/CategoryController.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using SmartSaver.EntityFrameworkCore.Models;
 using SmartSaver.Domain.Repositories;
-using SmartSaver.Domain.ActionFilters;
+using SmartSaver.Domain.CustomAttributes;
 using SmartSaver.WebApi.RequestModels;
 using SmartSaver.WebApi.ResponseModels;
 using AutoMapper;
